# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BASE",
        "description": "Servicio API REST para la gestión y administración del sistema Automatizado de riego por goteo",
        "contact": {
            "name": "SETI",
            "email": "cirojas@cenicana.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/water-sources": {
            "get": {
                "tags": [
                    "WaterSources"
                ],
                "summary": "Get list of water sources",
                "description": "Returns list of water sources",
                "operationId": "getWaterSourceList",
                "parameters": [
                    {
                        "name": "estate_id",
                        "in": "query",
                        "description": "The ID of the estate (hacienda)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "WaterSources"
                ],
                "summary": "Store new water source",
                "description": "Returns water source data",
                "operationId": "storeWaterSource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WaterSource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaterSource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/water-sources/{id}": {
            "get": {
                "tags": [
                    "WaterSources"
                ],
                "summary": "Get water source information",
                "description": "Returns water source data",
                "operationId": "getWaterSourcesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Water source id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/water-source-types": {
            "get": {
                "tags": [
                    "WaterSourceTypes"
                ],
                "summary": "Get list of water source types",
                "description": "Returns list of water source types",
                "operationId": "getWaterSourceTypesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WaterSourceType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WaterSource": {
                "title": "Water Source model",
                "description": "WaterSource model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the Water Source",
                        "type": "string",
                        "example": "Rio Cauca"
                    },
                    "capacity": {
                        "title": "Capacity",
                        "description": "Capacity of the Water Source / M3 para Reservorio , Rio and Pozo M3/h",
                        "type": "number",
                        "format": "float",
                        "example": "100.0"
                    },
                    "uptake_time": {
                        "title": "UptakeTime",
                        "description": "Uptake Timeof the Water Source in hours",
                        "type": "integer",
                        "format": "int64",
                        "example": "16"
                    },
                    "type_id": {
                        "title": "TypeId",
                        "description": "The ID of the water source Type: (Rio, Pozo, Reservorio)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "estate_id": {
                        "title": "EstateId",
                        "description": "The ID of the estate where belongs the water source",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "WaterSource"
                }
            },
            "WaterSourceType": {
                "title": "Water Source Type model",
                "description": "WaterSourceType model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "name": {
                        "title": "Name",
                        "description": "tipo de fuente de agua",
                        "type": "string",
                        "example": "Río"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "WaterSourceType"
                }
            },
            "EntityJsonResource": {
                "title": "EntityJsonResource",
                "description": "A Json wrapper resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EntityJsonResource"
                }
            }
        },
        "securitySchemes": {
            "jwt_key": {
                "type": "apiKey",
                "description": "Debe insertar el token generado en el proceso de autenticación",
                "name": "Authorization",
                "in": "header",
                "template": "Bearer {}"
            }
        }
    },
...
}